major optimizations and fixes with new features.

+++ src/models/address.js
@@ -12,30 +12,23 @@ class Address {
     }
 
     findAll = async userId => {
-        let sql = `SELECT a.id, a.user_id, c.name AS city, d.name AS district, a.title, a.address, a.last_update_date FROM ${this.tableName} a 
+        let sql = `SELECT a.id, c.name AS city, d.name AS district, a.title, a.address, a.last_update_date FROM ${this.tableName} a 
         JOIN districts d ON a.district_id = d.id JOIN cities c ON d.city_id = c.id WHERE a.user_id = ?`;
 
         return await query(sql, [userId]);
     }
 
-    findById = async (userId, id) => {
-        let sql = `SELECT a.id, a.user_id, c.name AS city, d.name AS district, a.title, a.address, a.last_update_date FROM ${this.tableName} a 
-        JOIN districts d ON a.district_id = d.id JOIN cities c ON d.city_id = c.id WHERE a.user_id = ? AND a.id = ?`;
+    update = async (userId, id, body) => {
+        const { keySet, values } = parseQueryParamsWithPlaceHolders(body, ', ');
+        let sql = `UPDATE ${this.tableName} SET ${keySet} WHERE user_id = ? AND id = ?`;
+
+        return await query(sql, [...values, userId, id]);//TODO: Set last_update_date
+    }
+
+    delete = async (userId, id) => {
+        let sql = `DELETE FROM ${this.tableName} WHERE user_id = ? AND id = ?`;
 
         return await query(sql, [userId, id]);
-    }
-
-    update = async (id, body) => {
-        const { keySet, values } = parseQueryParamsWithPlaceHolders(body, ', ');
-        let sql = `UPDATE ${this.tableName} SET ${keySet} WHERE id = ?`;
-
-        return await query(sql, [...values, id]);//TODO: Set last_update_date
-    }
-
-    delete = async id => {
-        let sql = `DELETE FROM ${this.tableName} WHERE id = ?`;
-
-        return await query(sql, [id]);
     }
 }
